requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
nodeenv==1.9.1
pyright==1.1.381
redis==5.0.8
sqlparse==0.5.1
